chore: address CodeRabbit feedback and remove dead code
- Update event system cleanup documentation to match implementation
- Fix type errors in fiber effect tags tests

Remove legacy code (~1,400 lines):
- Delete src/reconciler/ (replaced by Fiber architecture)
- Delete src/dom-renderer/ (no longer used)
- Delete tests/MiniReact.reconciler.test.ts (obsolete)
- Delete src/hooks/index.ts.backup (outdated backup)

Consolidate duplicated code:
- Refactor src/events/eventSystem.ts to import types from types.ts
- Merge src/hooks/fiberHooksImpl.ts into index.ts
- Remove ~150 lines of duplicate type definitions

Result: Cleaner codebase, zero breaking changes, all 520 tests passing
